backend/requirements.txt
asgiref
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT
pytz
sqlparse
yapf
psycopg2-binary==2.9.9
python-dotenv==1.0.0
whitenoise
gunicorn
whitenoise[brotli]
Faker
Pillow==10.1.0
cryptography==41.0.7
requests==2.31.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
celery==5.3.4
python-decouple==3.8
